+++ CHANGELOG.md
@@ -7,7 +7,21 @@ More expansive patch notes and explanations may be found in the specific [pathfi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.5.5] - 2023-05-18
+
+### Added
+
+- `cairo-lang` upgraded to 0.11.1.1
+
+### Fixed
+
+- RPC emits connection logs and warnings
+- Fee estimate mismatch between gateway and pathfinder
+  - Gateway uses a new gas price sampling algorithm which was incompatible with pathfinders.
+- Fee estimate returns error when submitting Cairo 1.0.0-rc0 classes.
+- Historic L1 handler transactions are served as Invoke V0
+  - Older databases contain L1 handler transactions from before L1 handler was a specific transaction type. These were
+    stored as Invoke V0. These are now correctly identified as being L1 Handler transactions.
 
 ### Fixed